requirements.txt
# -------------------------
# Core deep learning
# -------------------------
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# -------------------------
# Audio processing
# -------------------------
soundfile>=0.12.1
librosa>=0.10.1

# -------------------------
# Computer vision
# -------------------------
Pillow>=10.0.0
opencv-python>=4.8.0

# -------------------------
# Data handling
# -------------------------
numpy>=1.24.0
pandas>=2.0.0

# -------------------------
# Experiment utilities
# -------------------------
matplotlib>=3.8.0
tqdm>=4.66.0

# -------------------------
# Dataset & tooling
# -------------------------
yt-dlp>=2024.3.10